+++ packages/api/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+### Added
+
+- Cerbos v0.54 updates ([#1492](https://github.com/cerbos/cerbos-sdk-javascript/pull/1492))
+
+  - `cerbos.engine.v1.CheckOutput.evaluationErrors`
+  - `cerbos.engine.v1.EvaluationError`
+  - `cerbos.engine.v1.PlanResourcesOutput.evaluationErrors`
+
 ### Changed
 
 - Bump dependency on [@bufbuild/protobuf] to 2.12.1 ([#1478](https://github.com/cerbos/cerbos-sdk-javascript/pull/1478))